No soldier shields for murderers
An unintended consequence of the removal of faction loss and shield rank
on NPC soldiers is that murderers may have max soldier shields full time
with no consequences.

Unfortunately, murderers are already attackable by anyone, so they
receive soldier shield benefits without the soldier shield risks.

This is clearly backwards and inappropriate, creating a situation where
PKs are actually much stronger than hunters, and the proof is in the
pudding (every PK has a full time soldier shield).

This commit changes soldier shields so that murderers may not wield them
or keep them. This is only fair, since murderers cannot pay the risk
costs associated with a soldier shield, so neither should they receive
the benefits.