Royal Hunters / PK group size limits
If 3 or more outlaws or murderers gather in the same combat-enabled
area, royal hunters of the realm will appear and attack them all.

Royal hunters of the realm are protected by the decree of the Princess,
and cannot be harmed. They will kill murderers and outlaws if they do
not surrender and allow themselves to be captured.