Event Fixes
- Prisms no longer reset while on the ground.

- Due to constant griefing, the Nexus node has been given a new
alternative initiation. The spell Shatterlock can now manually open the
portal if the prism is used within the stonehenge.

- Shatterlock no longer functions with guild halls. Instead, it no
longer turns you orange, now has 1 hour base cast time, and will be used
to open (or force-start) many different kinds of events. The first event
of its kind is the alternative access to the Nexus node, not involving
tokens.

The 1 hour base cast time can be drastically reduced by using 5 or
7-sided prisms, and by increasing caster spellpower.

- Prism spells can no longer have their focus disrupted by anything
except resting, changing rooms, or running out of mana. This means
casters can fight to defend the prism. Logging all the casters is the
only way for an enemy to break the spell.