Added a chisel to remove gems with
The Jeweller's Chisel is a 'gem' wand that will pry gems out of an armor
for the player. He must be holding the armor but not wearing it.

The chisel will shatter after use, so that players aren't mad switching
gems during combat or something silly. Herbutte sells chisels for a
hefty fee, so removing gems will not be a casual affair.

Gems can still be removed from broken armors.